fetch: automatically trust keys
Code for trust moved to common/fetch_keys.go for sharing between
commands 'rkt trust' and other fetch commands.

The keys are only trusted automatically if they are fetched via https
and the images are fetched via the discovery protocol.
